feat isJoinableDateChallenge in ChallengeValidationService :  참여가능한 날짜의 챌린지인가? 확인을 위함.

+++ src/main/java/projectbuildup/mivv/domain/challenge/repository/ChallengeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import projectbuildup.mivv.domain.challenge.entity.Challenge;
 @Repository
-public interface ChallengeRepository extends JpaRepository<Challenge, Long> {
+public interface ChallengeRepository extends JpaRepository<Challenge, Integer> {
 }
